flaskrfid/requirements.txt
# Flask and core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database
pymongo==4.6.0
mongoengine==0.27.0

# RFID and NFC
pyscard==2.0.7
nfcpy==1.0.4

# Security
bcrypt==4.1.2
cryptography==41.0.8

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.2
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0
colorama==0.4.6

# Development
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# HTTP client for API calls
requests==2.31.0

# Internationalization
Flask-Babel==4.0.0